Move cckrg under surrogate_models
