+++ smt/applications/__init__.py
@@ -8,7 +8,7 @@ from .mfkplsk import MFKPLSK
 from .moe import MOE, MOESurrogateModel
 from .vfm import VFM
 from .podi import PODI, SubspacesInterpolation
-from .cckrg import CoopCompKRG
+from smt.surrogate_models.cckrg import CoopCompKRG
 from .mixed_integer import (
     MixedIntegerSamplingMethod,
     MixedIntegerSurrogateModel,
